api & jwt authentication added

+++ controllers/api/v1/posts_api.js
@@ -24,7 +24,7 @@ module.exports.destroy = async function(req, res){
     
     try{
         let post = await Post.findById(req.params.id);
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
@@ -32,9 +32,11 @@ module.exports.destroy = async function(req, res){
             return res.json(200, {
                 message: 'Posts & comments deleted'
             });
-        // }else{
-        //     return res.redirect('back');
-        // }
+        }else{
+            return res.json(401, {
+                message: 'You cannot delete this post!'
+            });
+        }
     }catch(err){
         return res.json(500, {
             message: 'Internal server error'
